trustchain_ig/requirements.txt
mcp>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
httpx>=0.24.0
python-json-logger>=2.0.0
prometheus-client>=0.17.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
cryptography>=41.0.0
aiosqlite>=0.19.0
asyncpg>=0.28.0